make utils.py, write todos
